+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/JodatimeDurationTypeConverter.scala
@@ -5,7 +5,7 @@ import com.google.inject.spi.TypeConverter
 import com.twitter.util.{Duration => TwitterDuration}
 import org.joda.time.Duration
 
-private[app] object JodatimeDurationTypeConvertor extends TypeConverter {
+private[app] object JodatimeDurationTypeConverter extends TypeConverter {
   def convert(value: String, toType: TypeLiteral[_]): Duration = {
     val twitterDuration = TwitterDuration.parse(value)
     Duration.millis(twitterDuration.inMillis)