+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/TwitterDurationTypeConverter.scala
@@ -4,7 +4,7 @@ import com.google.inject.TypeLiteral
 import com.google.inject.spi.TypeConverter
 import com.twitter.util.{Duration => TwitterDuration}
 
-private[app] object TwitterDurationTypeConvertor extends TypeConverter {
+private[app] object TwitterDurationTypeConverter extends TypeConverter {
   def convert(value: String, toType: TypeLiteral[_]): TwitterDuration = {
     TwitterDuration.parse(value)
   }